requirements.txt
Django==2.2.7
djangorestframework==3.10.3